feat: Placeholder `usePortals`
